Move lib/extensions to lib/business_time/core_ext
