+++ lib/business_time.rb
@@ -1,10 +1,12 @@
 require 'thread'
-require "active_support"
+require 'active_support'
 require 'active_support/time'
-require "time"
-require "business_time/config"
-require "business_time/business_hours"
-require "business_time/business_days"
-require "extensions/date"
-require "extensions/time"
-require "extensions/fixnum"
+require 'time'
+
+require 'business_time/config'
+require 'business_time/business_hours'
+require 'business_time/business_days'
+require 'business_time/core_ext/date'
+require 'business_time/core_ext/time'
+require 'business_time/core_ext/fixnum'
+
