+++ lib/business_time/core_ext/fixnum.rb
