+++ lib/business_time/core_ext/time.rb
